requirements.txt
streamlit
numpy
joblib
pillow
sklearn